# Optimized fetching of shapes and tracks