--- a/changelog.d/20250704_141107_sekachev.bs_optimized_attributes.md
+++ b/changelog.d/20250704_141107_sekachev.bs_optimized_attributes.md
@@ -1,5 +1,5 @@
 ### Fixed
 
 - Optimized database queries for attribute value removal.
- Replaced inefficient queries that caused sequential scans of large tables,
- improving performance and scalability (<https://github.com/cvat-ai/cvat/pull/9606>)
+    Replaced inefficient queries that caused sequential scans of large tables,
+    improving performance and scalability (<https://github.com/cvat-ai/cvat/pull/9606>)