--- a/changelog.d/20250707_060201_sekachev.bs_optimized_attributes_2.md
+++ b/changelog.d/20250707_060201_sekachev.bs_optimized_attributes_2.md
@@ -1,0 +1,4 @@
+### Fixed
+
+- Optimized query performance of retrieving shapes/tags from the database
+  (<https://github.com/cvat-ai/cvat/pull/9609>)