--- a/cvat/apps/consensus/merging_manager.py
+++ b/cvat/apps/consensus/merging_manager.py
@@ -127,7 +127,7 @@ class _TaskMerger:
         # imported annotations
         clear_annotations_in_jobs(
             [parent_job_id],
-            (self._task.project or self._task)
+            attr_ids=(self._task.project or self._task)
             .get_attributes(only_parent=False)
             .values_list("id", flat=True),
         )