<!-- Raise an issue to propose your change (https://github.com/cvat-ai/cvat/issues).
It helps to avoid duplication of efforts from multiple independent contributors.
Discuss your ideas with maintainers to be sure that changes will be approved and merged.
Read the [Contribution guide](https://docs.cvat.ai/docs/contributing/). -->

<!-- Provide a general summary of your changes in the Title above -->

Depends on #9606

### Motivation and context
Implementation before generates CURSOR with the following query:

```
SELECT "engine_labeledshape"."id",
       "engine_labeledshape"."label_id",
       "engine_labeledshape"."type",
       "engine_labeledshape"."frame",
       "engine_labeledshape"."group",
       "engine_labeledshape"."source",
       "engine_labeledshape"."occluded",
       "engine_labeledshape"."outside",
       "engine_labeledshape"."z_order",
       "engine_labeledshape"."rotation",
       "engine_labeledshape"."points",
       "engine_labeledshape"."parent_id",
       "engine_labeledshapeattributeval"."spec_id",
       "engine_labeledshapeattributeval"."value",
       "engine_labeledshapeattributeval"."id"
FROM "engine_labeledshape"
LEFT OUTER JOIN "engine_labeledshapeattributeval" ON ("engine_labeledshape"."id" = "engine_labeledshapeattributeval"."shape_id")
WHERE "engine_labeledshape"."job_id" = 2639313
ORDER BY "engine_labeledshape"."frame" ASC
```

Which based on left join and leads to seq scan of the table:

```
Sort  (cost=3634774.78..3634841.38 rows=26641 width=234) (actual time=38497.664..38499.381 rows=12000 loops=1)
   Sort Key: engine_labeledshape.frame
   Sort Method: quicksort  Memory: 12572kB
   ->  Hash Right Join  (cost=30555.68..3632816.49 rows=26641 width=234) (actual time=13510.315..38492.097 rows=12000 loops=1)
         Hash Cond: (engine_labeledshapeattributeval.shape_id = engine_labeledshape.id)
         ->  Seq Scan on engine_labeledshapeattributeval  (cost=0.00..3134215.55 rows=178302955 width=29) (actual time=0.034..18147.832 rows=178338113 loops=1)
         ->  Hash  (cost=30222.67..30222.67 rows=26641 width=213) (actual time=2.713..2.714 rows=3000 loops=1)
               Buckets: 32768  Batches: 1  Memory Usage: 3220kB
               ->  Index Scan using engine_labeledshape_job_id_b7694c3a on engine_labeledshape  (cost=0.57..30222.67 rows=26641 width=213) (actual time=0.017..1.343 rows=3000 loops=1)
                     Index Cond: (job_id = 2639313)
 Planning Time: 0.247 ms
 Execution Time: 38499.927 ms
```

Instead, it is better to make two queries:


```
SELECT "engine_labeledshape"."id",
       "engine_labeledshape"."label_id",
       "engine_labeledshape"."type",
       "engine_labeledshape"."frame",
       "engine_labeledshape"."group",
       "engine_labeledshape"."source",
       "engine_labeledshape"."occluded",
       "engine_labeledshape"."outside",
       "engine_labeledshape"."z_order",
       "engine_labeledshape"."rotation",
       "engine_labeledshape"."points",
       "engine_labeledshape"."parent_id"
FROM "engine_labeledshape"
WHERE "engine_labeledshape"."job_id" = 2639313
ORDER BY "engine_labeledshape"."frame" ASC
```

with query plan:
```
Sort  (cost=32180.96..32247.56 rows=26641 width=213) (actual time=2.373..2.592 rows=3000 loops=1)
   Sort Key: frame
   Sort Method: quicksort  Memory: 3097kB
   ->  Index Scan using engine_labeledshape_job_id_b7694c3a on engine_labeledshape  (cost=0.57..30222.67 rows=26641 width=213) (actual time=0.031..1.344 rows=3000 loops=1)
         Index Cond: (job_id = 2639313)
 Planning Time: 0.093 ms
 Execution Time: 2.742 ms
```

AND

```
SELECT "engine_labeledshapeattributeval"."shape_id",
       "engine_labeledshapeattributeval"."spec_id",
       "engine_labeledshapeattributeval"."value",
       "engine_labeledshapeattributeval"."id"
FROM "engine_labeledshapeattributeval"
WHERE ("engine_labeledshapeattributeval"."shape_id" IN
         (SELECT U0."id"
          FROM "engine_labeledshape" U0
          WHERE U0."job_id" = 2639313)
       AND "engine_labeledshapeattributeval"."spec_id" IN
         (SELECT V0."id"
          FROM "engine_attributespec" V0
          WHERE V0."label_id" IN
              (SELECT U0."id"
               FROM "engine_label" U0
               WHERE U0."project_id" = 291299)))
```

with query plan:

```
Nested Loop  (cost=495.34..500816.96 rows=1 width=29) (actual time=50.782..181.421 rows=12000 loops=1)
   ->  Nested Loop  (cost=494.77..474328.18 rows=3408 width=29) (actual time=1.039..27.779 rows=63200 loops=1)
         ->  HashAggregate  (cost=494.20..494.32 rows=12 width=4) (actual time=0.041..0.063 rows=16 loops=1)
               Group Key: v0.id
               Batches: 1  Memory Usage: 24kB
               ->  Nested Loop  (cost=0.85..494.17 rows=12 width=4) (actual time=0.020..0.034 rows=16 loops=1)
                     ->  Index Scan using engine_label_project_id_7f02a656 on engine_label u0_1  (cost=0.43..31.96 rows=47 width=4) (actual time=0.010..0.011 rows=4 loops=1)
                           Index Cond: (project_id = 291299)
                     ->  Index Scan using engine_attributespec_label_id_274838ef on engine_attributespec v0  (cost=0.42..9.78 rows=5 width=8) (actual time=0.003..0.004 rows=4 loops=4)
                           Index Cond: (label_id = u0_1.id)
         ->  Index Scan using engine_labeledshapeattributeval_spec_id_144b73fa on engine_labeledshapeattributeval  (cost=0.57..39350.03 rows=13613 width=29) (actual time=0.127..1.213 rows=3950 loops=16)
               Index Cond: (spec_id = v0.id)
   ->  Index Scan using engine_labeledshape_pkey on engine_labeledshape u0  (cost=0.57..7.77 rows=1 width=8) (actual time=0.002..0.002 rows=0 loops=63200)
         Index Cond: (id = engine_labeledshapeattributeval.shape_id)
         Filter: (job_id = 2639313)
         Rows Removed by Filter: 1
 Planning Time: 5.272 ms
 Execution Time: 182.001 ms
```


Total execution time on the same data is about 200 times less.

### How has this been tested?
<!-- Please describe in detail how you tested your changes.
Include details of your testing environment, and the tests you ran to
see how your change affects other areas of the code, etc. -->


### Checklist
<!-- Go over all the following points, and put an `x` in all the boxes that apply.
If an item isn't applicable for some reason, then ~~explicitly strikethrough~~ the whole
line. If you don't do that, GitHub will show incorrect progress for the pull request.
If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I submit my changes into the `develop` branch
- [ ] I have created a changelog fragment <!-- see top comment in CHANGELOG.md -->
- [ ] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] I have linked related issues (see [GitHub docs](
  https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword))

### License

- [x] I submit _my code changes_ under the same [MIT License](
  https://github.com/cvat-ai/cvat/blob/develop/LICENSE) that covers the project.
  Feel free to contact the maintainers if that's a concern.