@zhiltsov-max 2025-07-07
FYI: is you select the dependency branch as the target branch, the diff will only show the current PR changes.