@bsekachev 2025-07-07
Right now I am working on another solution with indexes because of some found inefficient planning on postgres side in rare cases